+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- new command `Add-FeatureState` for feature styling by state.
+
 ## [0.2.0] - 2023-10-19
 
 ### Added